Fix overflow for threads*particles

+++ platforms/cuda/src/CudaKernels.cpp
@@ -139,8 +139,8 @@ void CudaUpdateStateDataKernel::getPositions(ContextImpl& context, vector<Vec3>&
         Vec3 boxVectors[3];
         cu.getPeriodicBoxVectors(boxVectors[0], boxVectors[1], boxVectors[2]);
         int numThreads = threads.getNumThreads();
-        int start = threadIndex*numParticles/numThreads;
-        int end = (threadIndex+1)*numParticles/numThreads;
+        int start = threadIndex*(long long)numParticles/numThreads;
+        int end = (threadIndex+1)*(long long)numParticles/numThreads;
         if (cu.getUseDoublePrecision()) {
             double4* posq = (double4*) cu.getPinnedBuffer();
             for (int i = start; i < end; ++i) {
